// web/src/routes/_user/household/$householdId/{-$viewUserId}/categories/__generated__/CategoryIdQuery.graphql.ts
export type TransactionWhereInput = {
  and?: ReadonlyArray<TransactionWhereInput> | null | undefined;
  categoryID?: string | null | undefined;
  categoryIDIn?: ReadonlyArray<string> | null | undefined;
  categoryIDNEQ?: string | null | undefined;
  categoryIDNotIn?: ReadonlyArray<string> | null | undefined;
  createTime?: any | null | undefined;
  createTimeGT?: any | null | undefined;
  createTimeGTE?: any | null | undefined;
  createTimeIn?: ReadonlyArray<any> | null | undefined;
  createTimeLT?: any | null | undefined;
  createTimeLTE?: any | null | undefined;
  createTimeNEQ?: any | null | undefined;
  createTimeNotIn?: ReadonlyArray<any> | null | undefined;
  datetime?: any | null | undefined;
  datetimeGT?: any | null | undefined;
  datetimeGTE?: any | null | undefined;
  datetimeIn?: ReadonlyArray<any> | null | undefined;
  datetimeLT?: any | null | undefined;
  datetimeLTE?: any | null | undefined;
  datetimeNEQ?: any | null | undefined;
  datetimeNotIn?: ReadonlyArray<any> | null | undefined;
  description?: string | null | undefined;
  descriptionContains?: string | null | undefined;
  descriptionContainsFold?: string | null | undefined;
  descriptionEqualFold?: string | null | undefined;
  descriptionGT?: string | null | undefined;
  descriptionGTE?: string | null | undefined;
  descriptionHasPrefix?: string | null | undefined;
  descriptionHasSuffix?: string | null | undefined;
  descriptionIn?: ReadonlyArray<string> | null | undefined;
  descriptionIsNil?: boolean | null | undefined;
  descriptionLT?: string | null | undefined;
  descriptionLTE?: string | null | undefined;
  descriptionNEQ?: string | null | undefined;
  descriptionNotIn?: ReadonlyArray<string> | null | undefined;
  descriptionNotNil?: boolean | null | undefined;
  excludeFromReports?: boolean | null | undefined;
  excludeFromReportsNEQ?: boolean | null | undefined;
  hasCategory?: boolean | null | undefined;
  hasCategoryWith?: ReadonlyArray<TransactionCategoryWhereInput> | null | undefined;
  hasHousehold?: boolean | null | undefined;
  hasHouseholdWith?: ReadonlyArray<HouseholdWhereInput> | null | undefined;
  hasInvestmentLots?: boolean | null | undefined;
  hasInvestmentLotsWith?: ReadonlyArray<InvestmentLotWhereInput> | null | undefined;
  hasTransactionEntries?: boolean | null | undefined;
  hasTransactionEntriesWith?: ReadonlyArray<TransactionEntryWhereInput> | null | undefined;
  hasUser?: boolean | null | undefined;
  hasUserWith?: ReadonlyArray<UserWhereInput> | null | undefined;
  householdID?: string | null | undefined;
  householdIDIn?: ReadonlyArray<string> | null | undefined;
  householdIDNEQ?: string | null | undefined;
  householdIDNotIn?: ReadonlyArray<string> | null | undefined;
  id?: string | null | undefined;
  idGT?: string | null | undefined;
  idGTE?: string | null | undefined;
  idIn?: ReadonlyArray<string> | null | undefined;
  idLT?: string | null | undefined;
  idLTE?: string | null | undefined;
  idNEQ?: string | null | undefined;
  idNotIn?: ReadonlyArray<string> | null | undefined;
  not?: TransactionWhereInput | null | undefined;
  or?: ReadonlyArray<TransactionWhereInput> | null | undefined;
  updateTime?: any | null | undefined;
  updateTimeGT?: any | null | undefined;
  updateTimeGTE?: any | null | undefined;
  updateTimeIn?: ReadonlyArray<any> | null | undefined;
  updateTimeLT?: any | null | undefined;
  updateTimeLTE?: any | null | undefined;
  updateTimeNEQ?: any | null | undefined;
  updateTimeNotIn?: ReadonlyArray<any> | null | undefined;
  userID?: string | null | undefined;
  userIDIn?: ReadonlyArray<string> | null | undefined;
  userIDNEQ?: string | null | undefined;
  userIDNotIn?: ReadonlyArray<string> | null | undefined;
};
// ...
export type CategoryIdQuery$variables = {
  endDate: any;
  id: string;
  startDate: any;
  where?: TransactionWhereInput | null | undefined;
};
export type CategoryIdQuery$data = {
  readonly household: {
    readonly financialReport: {
      readonly " $fragmentSpreads": FragmentRefs<"categoryCardFinancialReportFragment">;
    };
    readonly " $fragmentSpreads": FragmentRefs<"transactionsListFragment">;
  };
  readonly node: {
    readonly id?: string;
    readonly " $fragmentSpreads": FragmentRefs<"categoryCardCategoryFragment" | "editCategoryFragment">;
  } | null | undefined;
};
export type CategoryIdQuery = {
  response: CategoryIdQuery$data;
  variables: CategoryIdQuery$variables;
};